refactor(app): Make toggle components reusable
Group toggle button, labeled toggle, and toggle info for resuse in app and robot settings
