--- app/src/components/ToggleButton/styles.css
@@ -1,10 +1,0 @@
-@import '@opentrons/components';
-
-.toggled_on {
-  fill: #0ec9eb;
-  color: #0ec9eb;
-}
-
-.toggled_off {
-  fill: var(--c-dark-gray);
-}
