+++ app/src/components/analytics-settings/AnalyticsSettingsModal.js
@@ -5,7 +5,6 @@ import {connect} from 'react-redux'
 import {getAnalyticsSeen, setAnalyticsSeen} from '../../analytics'
 
 import {Modal} from '@opentrons/components'
-import ModalTitle from './ModalTitle'
 import ModalButton from './ModalButton'
 import AnalyticsToggle from './AnalyticsToggle'
 import AnalyticsInfo from './AnalyticsInfo'
@@ -36,10 +35,7 @@ function AnalyticsSettingsModal (props: Props) {
   const {setSeen} = props
 
   return (
-    <Modal onCloseClick={setSeen} alertOverlay>
-      <ModalTitle>
-        {TITLE}
-      </ModalTitle>
+    <Modal onCloseClick={setSeen} heading={TITLE} alertOverlay>
       <AnalyticsToggle />
       <AnalyticsInfo />
       <ModalButton onClick={setSeen}>
