+++ app/src/components/toggles/index.js
@@ -1,0 +1,6 @@
+// @flow
+import ToggleButton from './ToggleButton'
+import LabeledToggle from './LabeledToggle'
+import ToggleInfo from './ToggleInfo'
+
+export {ToggleButton, LabeledToggle, ToggleInfo}
