+++ app/src/components/toggles/styles.css
@@ -1,0 +1,49 @@
+@import '@opentrons/components';
+
+/* Toggle Button */
+.toggled_on {
+  fill: #0ec9eb;
+  color: #0ec9eb;
+}
+
+.toggled_off {
+  fill: var(--c-dark-gray);
+}
+
+/* Labeled Toggle */
+.labeled_toggle {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  width: 100%;
+  margin-top: 0.5rem;
+}
+
+.labeled_toggle_label {
+  @apply --font-body-1-dark;
+
+  font-weight: var(--fw-semibold);
+}
+
+.labeled_toggle_button {
+  flex: none;
+  width: 2rem;
+  margin-right: 1rem;
+  padding: 0;
+
+  &:hover {
+    background-color: transparent;
+  }
+}
+
+/* Toggle Info */
+.toggle_info {
+  @apply --font-body-1-dark;
+
+  max-width: 25rem;
+
+  & > p {
+    margin: 1rem 0;
+    line-height: 1.5;
+  }
+}
