+++ components/src/modals/Modal.js
@@ -7,6 +7,8 @@ import styles from './modals.css'
 type ModalProps = {
   /** handler to close the modal (attached to `Overlay` onClick) */
   onCloseClick?: (event: SyntheticEvent<>) => mixed,
+  /** Optional styled heading **/
+  heading?: string,
   /** modal contents */
   children: React.Node,
   /** classes to apply */
@@ -22,11 +24,14 @@ type ModalProps = {
  * with a dark overlay and displays `children` as its contents in a white box
  */
 export default function Modal (props: ModalProps) {
-  const {contentsClassName, alertOverlay, onCloseClick} = props
+  const {contentsClassName, alertOverlay, onCloseClick, heading} = props
   return (
     <div className={cx(styles.modal, props.className, {[styles.alert_modal]: alertOverlay})} >
       <Overlay onClick={onCloseClick} alertOverlay={alertOverlay}/>
        <div className={cx(styles.modal_contents, contentsClassName)}>
+         {heading && (
+           <h3 className={styles.modal_page_heading}>{heading}</h3>
+         )}
         {props.children}
       </div>
     </div>
